refactor: extract OpenAPI request-body handling into utils module
Move the stateless request-body helpers introduced for #1078 out of
clients/openapi.ts (~280 lines) into utils/openApiRequestBody.ts so the
client keeps connection/security/session concerns and the serialization
logic is unit-testable in isolation. Pure code move — no behavior change;
adds direct unit tests for content-type priority, circular-schema
tolerance, multipart framing/header sanitization, and form encoding.
